Work towards converting sbu.py
Got most of the current Python functionality reproduced in C++ (which
can be called conveniently in bash scripts, etc.).  Now just need to
build the unique vectors, and it'll pretty much be ready.

Note: take note of BABEL_DATADIR when I eventually reset the atoms list.
Since data.cpp calls Init(), there might be a way to add an override
option for the data source (or maybe append another directory?).