Start hMOF decomposition testing
Decompose hMOFs to see if the linkers identified by OpenBabel match the
hMOF specification.  Based on very preliminary results, this test
appears like it will take a day or less of computation to perform.  Next
steps are diagnosing some of the linkers, extracting nodes, treating
functionalization, and verifying the entire database to look for trouble
spots.