Fix oxygen and water bonding in NU-1000
When using SMILES as a basis for filtering out types of building
blocks, we need to be careful to redetect the bond types first for
consistency (purpose of resetBonds(OBMol)).  There was also a hidden
problem involving the maximum valency of Zr: 6 was not enough, so I
upped it to 8 temporarily.  This should be addressed in the long run.