Export 1_565 UC bond flags
Calculate the unit cells between bonds.  This seems to work very well
in the limited tests exported from RingCIFs/.  Bonds are appropriately
assigned to the correct unit cell.  Overall, this code will be a very
useful diagnostic for inspecting bonds and visually understanding a
more "exact" representation of the OBMol state.  Implementation is
similar to the RCSR export code, saved in ManyNets:write_cif.py:103.

One unresolved bug is in the bond orders: central BDCs look reasonable,
but ones on the borders have the double bonds in the wrong spots (three
next to each other, etc.).  Two probable causes: the radical from
disconnecting bonds is messing up the procedure, and/or I missed PBC
somewhere in the bond detection (checking OBMol::NewPerceiveKekuleBonds
and a few other places briefly).