Fix segfault when combining nodes
Open Babel (and Materials Studio) would segfault if collapseSBUs was
applied to the nodes because the bonding accounting was incorrect.  It
turned out that the OBAtom bond indices needed to be updated.  Now that
this bug is corrected, there are a few things remaining before passing
the structures to Systre for topological identification.