Successfully export simple CGDs
The code still has bugs in certain cases (small unit cells, 2-connected
nodes aren't simplified, etc.), but for some "easy" cases
(super_nu1000.cif) the CGD will successfully work within Systre to
identify the topology (csq in this example).  Still need to automate
this process and fill in some of the missing features, but it's
certainly a step in the right direction.