Compare ToBACCo identities against MOFFLES
Refactor a little bit of the MOFFLES script so we can easily compare
the generated node, linker, and topology identification generated from
MOFFLES as compared with the linkers established in the ToBACCo file
name conventions.  I still need to analyze the data saved in the
Notebooks/ directory tomorrow to make sense of the high failure rate:
likely differences in radicals on the linkers.