Use local data instead of global OB files
We do not want to use the global data for Open Babel saved in
BABEL_DATADIR, because that one is configured for general molecules
(and other programs saved on the user's computer).  The MOFFLES code
will require custom adjustments, particularly in the maxbonds of
elements.txt, to handle the unique bonding environments encountered by
metals and oxygens within metal nodes.