Generalize code for hMOFs
Apply the object-oriented code from ToBACCo to the hMOFs as well.  The
Python code will be a good base for various tests.  A few changes were
required to the HMOF data file to account for the new (more complete)
description of node SMILES and a little bit of updated linker info,
although the file is still far from complete.  The Python code also
cannot detect which constructor is necessary (yet).