Formalize known linkers and measure timing
Added a few MOFs of interest from the earlier RingCIFs/ work as
benchmarks: these are already (mostly) correct, or have a known
deficiency that we're actively working towards correcting.  This
commit also added timing information to the `test_cif` method, which
will identify MOFs that the algorithm has difficulty with, as well as
measuring improvements or regressions on the code performance.