Skip kekulization of whole MOF
Massively improved the performance of the code (~7.5x for some MOFs,
>5x overall) by skipping the full detection of bond orders on the
entire input MOF.  Instead, just calculate the non-scalable
kekulization algorithm on the individual nodes and linkers.  This code
change introduced strange bonding in P1-IRMOF-1.cif, so the
topological definition is temporarily set to an artificial dia-f to
keep the verification testing clean.