Import Systre code from OB format
Repurpose my Systre code from the output format to be adapted to the
new connection-based output (where two-connected links are explicitly
defined).  Fixed whitespace inconsistencies in the code and ran the
full 13 MOF test on a slightly modified version of sbu.cpp (which
temporarily disabled the use of the X pseudoatoms).