Patch the nitrogen linker regression
Instead of reporting the nitrogen linker `[nH]` failure as a known
error, just generically fix it.  This will likely solve future
problems with functionalized nitrogen linkers of this type.  Also, the
purpose of the Python code is to generate the expected name based on
the crystal generator input parameters.  Thus, its purpose is to
sytematically generate the expected SMILES.  The result is our
objective, not actually what we're trying to validate.