Improve performance of copying OBMols
Earlier, the largest bottleneck in the code was the seemingly
innoculous copying of the OBMol class for the original MOF.  However,
this code implicitly called ring perception routines, etc., which did
not scale well to the whole MOF.  Avoid these using a new `copyMOF`
function, which overall improved the performance somewhat, but
especially for the largest structures.  Now, the greatest bottleneck
is caused by Systre nearly-crashing for some topologies.