Remove free solvents
Delete free solvent molecules from the MOF, which here are defined as
organic species not connected to any metals.  This partially works on
the CSD structure NADZEZ, but to complete the implementation, the
solvent removal will need to be effective on single atoms and/or loose
oxygen species.  In the spirit of the CoRE MOF databases, the code
also reports the desolvated structures.