Fix bug with simplifying one-connected nodes
The order of simplifying the MOF net will determine if the MOF's
state is reasonable.  We should collapseTwoConn after the other types
of simplifications to avoid the M1-X-L-X-M1 case, where
collapseTwoConn would simplify it to M1-X-X-M1 and M1-X-M1 without
considering that the two X connections are in the same direction and
redundant.

New test cases address this issue, though the Zr MOF should be "new"
(and not a Systre error).  For whatever reason, Systre is finding
collisions in the structure currently.