Defer exporting the node and linker information
Moved writing of the CIFs for nodes, linkers, original molecule, and
some other "helper" info files to the end of the script.  This will
allow the code for removing bound solvents (for ASR) to be placed
after other simplifications to the simplified_net, which is crucial
for removing them from the detected linkers.