Fix segfault when many pseudo atoms are used
The unrelaxed ToBaCCo CIF `ssa_sym_5_mc_2_sym_5_mc_2__.cif` was
causing a segfault when checking for the consistency of catenated
nets.  This in turn was caused by the seemingly innocent `GetFormula`
method in OBMol, which has not been updated for heavy elements.  This
is a pretty easy bug in the upstream Open Babel distribution, so I
will write a few test cases and submit it back.