Clean bonds in fully simplified net
Handled atoms on the +1 edges of unit cells to avoid problems with the
visual output in removed_two_conn_for_topology.cif.  The root cause
was a failure to wrap cartesian coordinates in collapseXX.  The legacy
calculation-by-hand has now been replaced with the general getMidpoint
code.  Note that long-term changes to `OBBond::GetPeriodicDirection`
may be necessary, but these can wait until we have a community
consensus for how to handle atoms on the edges of unit cells.