Fix tet stereochemistry for periodic mols
Continue adapting perception.cpp to account for periodicity in systems
with tetrahedral chiral centers.  After implementation, metal nodes
(with the exception of strange, porphyrinic ADARAA) are canonicalized
and no longer show duplicate entries!

In the stereo/perception.cpp code, there are still unresolved
references to GetVector, but these seem to be focused on the 2D cases
where periodicity would be unimportant.  PerceiveStereo calls the
corresponding StereoFromND based on dimensionality N, and it looks
like I've covered the important cases for 3D.