Update skeleton for simplifyLX
We want to simplify ToBaCCo MOFs containing oxalic acid.  This
implicit linker is currently a problem due to the angle criterion
in simplifyLX and a lack of the linker infrastructure in the Python
testing code.  Set up the code to add an alternative to the angle
threshold of 66 degrees for redundant LX bonds, because the angles are
highly sensitive in small linkers.