Detect redundant LX bonds by searching for chains
Reuse the isPeriodicChain code from MIL-47-family detection to help
simplify the LX bonds.  Earlier, I was using a maximum angle threshold
to differentiate metals in different unit cells.  This update uses a
better algorithm more closely aligned with the objective: we just want
to differentiate metals in different unit cells.  The simple way to do
that is detecting an M-X-L-X-M that forms a periodic chain (otherwise,
if it's redundant, it will be contained as it's own cluster).  As part
of this update, the code identified bound solvent on a plain,
hydrogen-terminated Zn4O(CO2)6 cluster, which passed the angle
criterion but should actually be classified as solvent (or linker).