Ignore hydrogen in molecular formula diff
We want to find the difference between the molecular formula between
two SMILES.  Later on, we exclude hydrogen because the number of
implicit atoms will be dependent on the calculated bond structure, so
also ignore it when comparing molecular formulas.  Also changed the
structure of `sym_5_on_12` to better reflect the SMILES from Open
Babel.