Generalize metal oxide detection
Oxygen-related species, such as hydroxyls and even peroxides, were
detected as parts of nodes by explicitly enumerating the expected
SMILES for each species.  Instead, let's simply look for disconnected
fragments that only contain oxygen and hydrogen, which is more
generalizable and fixes a few errors encountered when screening a
subset of the ToBaCCo MOFs.