Fix segfault with sym_3_mc_0
One of the metal nodes in ToBaCCo was crashing the sbu analysis code
in the simplifyLX routine due to a poor assumption that L-X will
always have a metal neighbor.  However, we're skipping implementation
of the node in the ToBaCCo database right now because (1) the topology
is going to be reported differently in this decomposition scheme, and
(2) the linker is rejected by rdkit's kekulizer due to the radical vs.
charge notation.