Make Systre cluster-friendly
One of our clusters needed more memory for Systre to avoid an error
about the heap size, so increase that to 1GB.  Also add code to
install the Systre jar file in the resources directory, unless it's
installed locally and the user wants to specify the files in
Python/extract_moffles.py.  Also keep the ignored directory that
contains external resources by force adding a .gitkeep file.