Parse Systre output on catenated topologies
Fixed crash on structures like CoRE MOF IGAGUU_clean, which contains
multiple nets with the same NEW topology.  In these cases, Systre
reports that it has already seen the topology in a previous component,
not stating that it's a completely new topology.  This fixes the
interpretation of the Systre output and avoids a false assert.