Attempt to add charges to N-containing linkers
Tried assigning charges to imidazolate and related linkers using
SMARTS matches.  Unfortunately, this failed due to the earlier
assignment of a radical to other atoms in the aromatic ring,
leading to an inconsistent molecular fragment.  This may easily be
fixed via changes in the upstream project to the source code, which
recalculates the atomic spin instead of perceiving and setting.