Ensure paddlewheels are consistently bonded
Detect paddlewheels using a SMARTS query (and exclude infinite MIL-
type rods by checkign for periodicity).  Then, draw a bond between the
two metal atoms, which ensures that they are lumped together as a
single node in the simplified net.  This commit will permit larger
screening of the GA hMOFs.