Refactor bound solvent removal into removeOneConn
Moved bound solvent removal from the main function into its own
subroutine and moved its invocation to within the simplification loop,
which enables it to take effect on axially-connected paddlewheels
(such as GA3001840).  Previously, the solvent was removed after the
simplification loop, so it was marked as a two-connected node too
late.  More work required to address topologies in these structures.