Ensure all MOFid outputs are .smi compatible
Output from Python/extract_moffles.py can be collected into a .smi
database, e.g. from screening the CoRE MOFs.  Change the notation for
some of the lines, such as "ERROR" (for C++ crashes) or "" (when MOF
components are not found in the CIF) to be a somewhat standard "*",
which both Open Babel and rdkit can parse.