Implement GA linkers 34-37 and transform two-conn. links
Added organic corners 34-37 to the database of known GA building
blocks, although linker 37 currently has a problem with forming the
C#N bond in a purchasable compound.  Also changed the logic of
collapseTwoConn to replace two-connected components with a connector
pseudo atom instead of removing it from the network entirely.  This
change in the algorithm prevents MOFs like h3001840 from obtaining the
wrong X-X bond directions.