Allow linker to only be pillar, and handle linker 2
A huge number of ParentGA MOFs (200) were being classified as having
the wrong formula because the pillar MOF rules assumed that the
primary or secondary linkers would be present in their carboxylate
form.  It turns out that they can also exclusively be the pillar,
allowing the other linker to take the carboxylate form.  One of the
linkers (L2) was defined with the wrong nitrogenated form.