Fix inconsistency in PW bond assignment
Work on paddlewheels in July used a SMARTS-derived heuristic to
differentiate paddlewheels from periodic rods (and small unit cells),
but reformation of metal-metal paddlewheel bonds in `resetBonds` was
set to use a distance method.  Reworking the reset code to use a
minimum distance removes this (only) free parameter in the MOFid code.

Also added a regression test case for **pcu** GA hMOFs in small unit
cells, which can be misidentified as the **xah** topology.