Remove lattice helper functions from upstream OBMol
My previous iteration of periodic boundary conditions in OBMol added
an extra copy of OBUnitCell* to the OBMol object, which was
undesirable because (1) it clutters the Open Babel API, and (2) leads
to consistency problems.  See also how the code in sbu.cpp is now more
straightforward now that we don't have to worry about which copy of
OBUnitCell we're using for a MOF/molecule.  Also committing the OB
part of these changes to my upstream branch in progress towards
getting the code into the main Open Babel trunk.