Reset element.txt to remove max valence hacks
Restored the maximum valence of oxygen to 2 within element.txt, which
will help deprecate this file when the upstream Open Babel code is
merged with the newest development version.  In the process, I found a
bug in the super-valence oxygen bond assignment.  Restoring the oxygen
valence also fixed the notation of hydroxyls/waters in the SMILES,
requiring some patches to the known MOF SMILES.

Up next, there's a regression in one of the Zn paddlewheel hMOFs.
The reference nodes in the ToBaCCo/hMOF/GA MOF json files will also
need updates.