Adapt (C)Makefiles to compile code in Cygwin
Was having problems getting the projects to compile after updating the
upstream Open Babel project, so tried a number of adjustments
including building in a separate directory from the source code.  In
the end, it seems likely that the main incompatibility was a missing
DLL, so a future commit should roll back most of the changes from this
commit.