Use the same BABEL_DATADIR for Python and web
The Python test suite was raising several errors now that the upstream
Open Babel code has been updated, because some of the required 2.4.1
input files (like element.txt) were deprecated in the development
version of the code.  Now the Python path is consistent with the path
in config_sbu.h.