Switch most of the Python dependencies to C++
Replaced most of the calls to the pybel Python library with calls to
the obabel executable or a custom transformer code.  These changes
were made to ensure that the SMILES handling is 100% consistent using
the exact same version of Open Babel, which also has the side effect
of simplifying installation prerequisites on a cluster.  The downside
is that the code is now slower due to the overhead of forking.