Export SMILES to restore an operational test suite
I checked the output for unexpected errors or warnings as a result of
the new PseudoAtom/connection handling.  super_hMOF_22242 has many new
problems with unwrapFragmentUC.  The COD structure also triggers
several errors.  Otherwise, the test suite still runs in ~2 minutes
and is successful for 17 of the 27 MOFs, not bad considering that my
code is still missing some features.  Looking through the output
SMILES, most of these are improper calls and/or missing data for
resetBonds, plus needing to remove bound solvent as currently tested.