Write a first attempt to remove bound solvents
Realized that the one-coordinated simplifications can be much simpler
and streamlined now that connections are handled explicitly, so the
steps no longer have as fragile of a sequence.  The code needs to be
cleaned up, but at least preliminary testing shows a reduction from 10
errors to 6 in the test suite: not detecting the topology for
h1003830, wrong paddlewheels for 3 MOFs, and missing rod handling code
for two MIL MOFs.  Then the test suite will be operational again.