Fix bugs with PW in the test suite
The OBMol exporter for VirtualMols was not properly copying
paddlewheel detection flags to the new OBMol.  The test statement
was checking the newly formed atom copy for paddlewheel flags instead
of the original atom, so paddlewheels were never flagged.

This commit reduces the `make test` errors to three: two for MIL-47
structures plus hMOF-1003830.