Customize output folder from Test/
Changed the default output folder to Output/ and made both the sbu.cpp and Python code more flexible for different folder names.  This commit will make it possible to run multiple MOFid jobs currently in parallel using the same compiled folder (provided that the intermediate output folders are specified in the submit.job), which will be useful for code validation tests on MOF databases.