Avoid collapsing a 1-c metal into a 1-c linker
This commit will partially handle the case of metal-containing solvents or isolated metal nodes.  One test case was failing from the previous commit, likely caused by an inconsistent order of operations when attempting to simplify the Zn4O_cluster_with_hydrogen, which is a nonperiodic MOF-5 node with attached carboxylate.