Fix sub-process calls on Linux
The Python/check_mof_linkers.py code was failing on a Linux cluster because obabel was unable to parse the quote in the command line argument.  These quotes are essential on Cygwin (and Bash) to avoid globbing but cause problems for some reason using the Python wrapper in Linux.

As part of the debugging process, I also replaced the easyprocess package with subprocess32, a compatibility layer to use Python 3.x subprocess calls within Python 2.7.  A few conventions like newlines from stdout were slightly different, but otherwise it was a fairly straightforward transition.