Export SBU's, removing bonds between points of extension
Reassigned points of extension as part of the linker instead of deleted atoms.  This approach is more natural for the rest of the topological assignment and connection code.  Then, generating the SBU's consists of gathering the nodes plus points of extension, minus bonds between adjacent SBU's.

This commit also reverts changes to topology.h/cpp: the method DeleteAtomKeepConns is no longer necessary, and it's unclear if overlaying multiple connection PA's would cause problems down the road.  I'm removing it for maintainability, since we can always retrieve it later from Git.