Avoid inconsistent Systre output in nets with adjacent 2-c sites
If linker simplification is not complete, an organic building block could be composed of multiple adjacent 2-c pseudoatoms (e.g. the carbons in BDC if hydrogens are stripped away).  This is incompatible with the current version of Topology::WriteSystre(), so flag it as such instead of writing an inconsistent net.