Consider valence for assigning BP vs. EXT_CONN
Reworked external connections to consider them as an implicit bond to add to the explicit valence of an atom, transforming PA's containing external connections into branch points if they also had multiple explicit connections.  This notation sets the stage nicely for a demarcation between the explicit branch points to keep in the simplified net vs. types of branches to consider more like bonds.

Also fixed a bug in one of the loops where it improperly reused a variable from a previous loop.