Fix crash when simplifying a connector-only branch
The unwrapFragmentUC utility was not robust to an empty OBMol, leading to a segfault for structures like opt_tobacco/all/bor_sym_3_mc_0_sym_4_on_8_L_18.cif.  In the MappedMol, the organic building block would contain two 3-c rings connected by a ConnectionTable PA branch.  However, this mapped branch does not correspond to any real MOF atoms, so trying to simplify it leads to the zero-atom case.  This commit accounts for this situation.