Include carboxylates in the nodes, including the PoE
This commit reconsiders carboxylates as part of the node, even if binding end-on instead of chelated to two metals/oxygens.  I had formerly classified the point of extension (carboxylate carbon) as part of the linker, but that assignment is incompatible with an end-on binding, and I thought the concepts might be clearer if it's kept with the node.

This commit technically fixes the test suite but introduced at least one new bug in bond formation at the end of WriteSBUs, which is flooding the output with formBond errors for MOFs like MIL-47 and ABAVIJ.  In preliminary work, I think the bug is with the periodic rod simplification, because the carboxylates should still be simplified as one unit instead of three separate atoms.  There might also still be an inconsistency somewhere with PoE-PoE bonds and the original atoms forming unexpected clusters.