Properly simplify carboxylates in rod-like MOFs
Carboxylates are now simplified to single-PA node bridges when splitting up MIL-like infinite rods in CollapseNodes instead of using three discrete atoms.  The formBond bug from the previous commit was ultimately caused by an inconsistent accounting of nodes vs. node bridges.  We were iterating over the points of extension and their neighbors in WriteSBUs but had not considered the node bridges as part of the SBU's, even though they should be included.