Organize MOFidDeconstructor output into a new subdirectory
Cleaned up the Output/ directory by organizing the base organic/inorganic split into a NoSBU/ subdirectory, given that name because the algorithm does not have explicit SBU detection.  The output is now cleaner and easier to view.  We do not have to worry about the path to topology.cgd because the Python defaults have already been switched over to the SingleNode algorithm.