Denote alternate topological representations with a comma
This commit extends the MOFid format to allow multiple comma-separated topological representations, e.g. adding the all-node format representation when distinct from the default single-node.  Making this change requires a subtle adaptation to the topology verification, which now has a ToBaCCo MOF test case.  Overall, calculating both the single-node and all-node topologies increases the computation time of the test cases on my laptop from ~2 minutes to 3 minutes.  The names_to_tables.py converter is also aware of the possibility of multiple topologies, so it exports a processed base topology and a separate file for any additions.