Remove list_to_moffles.py
This code was used in the past to calculate the overlap between MOF databases by converting a list of ToBaCCo MOF filenames into a .smi containing their expected MOFid's.  Let's stick to check_mof_linkers.py for validation and the actual mofid.smi output for database comparison.  If there's a problem in the ToBaCCo structure, it's also likely present in the CoRE MOF version.