Export linker InChI's
The Output/NoSBU/inchi_linkers.txt file will be a useful diagnostic for understanding any strange behavior in the exported InChIKeys.  Once hashed, it's difficult to understand what the InChIKey says (other than a go/no-go if it's indexed online).  However, the original InChI can pasted into tools like ChemDraw for more insight about the structure that the InChI program is working with.