Fix MOF name regression in convert_smi_to_tables.py
Fixed bug where convert_smi_to_tables.py would include a trailing newline as part of the MOF name, so the name field of the TSV's would push the other fields onto a newline.  This behavior was caused by the change from line.split() in parse_moffles (not allowing spaces in names, but also treating the ending newline as a separator) to line.split(';') in parse_mofid.  Normalizing the input MOFid string to right-strip whitespace fixes this regression.