Update pip flags in README instructions
Added the `--user` flag to pip to avoid globally installing the MOFid package.  On computing clusters like NUIT quest, the user often does not have write access to the package directory, so this flag will avoid that pitfall.  The `-e` (or `--editable`) flag makes a reference to the MOFid directory instead of actually installing (copying?) the package.  That means if you make a change in mofid's Python/ directory, it will be applied to the installed mofid module without needing to reinstall.  As a side benefit, this change bypasses a minor annoyance where pip would take several seconds to install the package (perhaps it was looking for Python files elsewhere in `.`?).