Manually import openbabel PR to fix coordinate viz
Applied changes to openbabel/src/mcdlutil.cpp from openbabel/openbabel#1902 to fix the generation of 2D coordinates, which is essential to convert the SMILES and display molecules on the website.  This commit addresses the most important parts of #13 though it would still be good to fully merge the Open Babel version long-term.