Rename "NEW" topology flag as "UNKNOWN"
If Systre calculates a topology not reported in the RCSR, it prints "Structure is new for this run".  Using the word "NEW" in MOFid/MOFkey is misleading, because the topology may not necessarily be novel.  (It could be an improper decomposition of the MOF, topology found in another database, newer version of RCSR, etc.).  "UNKNOWN" is a more precise status and self-describing.