Implement a similar breakdown as the "standard algorithm"
Added a new folder that functionally serves the same purpose as the "standard algorithm," which consistently considers all metal atoms as isolated species.  Thus, the standard representation of MOF-5 yields the topology fff.  Note there may be subtle differences in implementation compared against other software (e.g. ToposPro), such as the bond assignment algorithm and the handling of 2-coordinated vertices in the simplified net.  This commit is mostly added as a demonstration for alternate algorithms.