Restore parsing of sbatch parameters
Unlike moab, slurm expects the job parameter flags (name, stderr/stdout locations, etc.) to be given before the .job file (because flags after that are passed as part of the command).  This commit fixes that reporting and also aggregates the folder_smiles_parts.tsv data file established as part of PR #20.