Report singleton MOFs in duplicates analysis
Writing a new file that does not filter families by `qty > 1`, giving a file with one line per header plus the "Number of duplicates families".  By adding a new command line argument to the `duplicates` flag, the duplicates and overlap analyses now require the same number of arguments, so there was some light rearrangement of the argparsing code.

The README now provides an example running the code over all three databases (CoRE, GA, ToBaCCo), which will be helpful for the overlaps analysis.