Explicitly exclude MOF building blocks from validation, per SI
Simplifying the workflow and discussion of the validation by outright including/excluding certain building blocks with justification.  That will be easier than the former approach, which had a mix of a few exclusions but mostly explanations on top of error codes.

In order to make this operational, I restored the missing building block SMILES to the tobacco_info.json data file.