Fix regression in Python parsing
As of commit f1839455, the C++ decomposition code differentiates between node and linker SMILES.  However, the Python parsing code now crashes when the C++ code encounters a segmentation fault or other program crash.  I thought the issue was with searching for `all_fragments[-1]`, but it turned out to be a variable that had not been renamed from `fragments` to `all_fragments`.  This commit fixes that bug and a related edge case.