Remove emscripten svg
I originally attempted to move the inline svg to an external file (the inline code makes SmartGit interpret the html file as binary instead of text for comparison reasons), but the namespaces and other details were not working.  Since an informative title is more useful long-term, anyway (and possibly a new logo), I removed it entirely.