Attempt to import bonds using _struct_conn
After implementing this feature, I found that cif-parser.js will not run the processConnections method to read this section of the CIF if we are reading in the atomic coordinates via atom_site_type and fract_x, etc.  Since they're fundamentally incompatible, we need to switch to a different file format if the bonding is important, unfortunately.