Update emscripten installation
Added more general commands to build the web interface.  Newer versions of emscripten require a special application/wasm MimeType on the HTTP Server configuration per https://www.w3.org/TR/wasm-web-api-1/ so this commit starts with simple asm.js code.  This commit was tested on the NUIT quest cluster, but more work is still needed to restore the performance of the website.