Work around performance bottleneck from strftime
According to a [blog](https://blog.magnum.graphics/announcements/new-emscripten-application-implementation/), newer versions of Emscripten contain a JS implementation of strftime as soon as C++ streams are introduced (several levels down into the STL implementation due to timers).  In turn, for the SBU code, the RegEx lines modifying EXPANSION_RULES then require ~45 seconds on a calculation that would otherwise complete in 15-25 seconds (including the JS topology impelmentation).

We do not actually use the results of strftime: it's just not optimized out of the code for some reason.  As a result, we can disable it with a sed command to restore the initial stub implementation, before the [upstream PR](https://github.com/emscripten-core/emscripten/pull/1424/files).